Update run-app.yaml
We want our action to build our project. To do this, navigate to the file .github/workflows/run-app.yaml. This is where we have define a workflow that runs a process called Basic Action that echos Hello, World!. It runs on the latest Ubuntu OS and runs everytime you push.

Replace the contents of the file with the code below